Fixed some stufff
I fixed an issue that let you get sheild faster from a sheild overtime hitboks by going inn and out of it.

Made a funksion sets sprinting to fase when the player is standing still. There is a issue with this.

Started to make a script that can select a weapon if the is a GO in the scene.